base_requirements.txt
absl-py==2.1.0
copulas==0.11.0
flax==0.8.5
matplotlib==3.9.1
ml_collections==0.1.1
optax==0.2.3
optuna==3.6.1
pandas==2.2.3
qiskit==1.1.1
scikit_learn==1.5.1
tabulate==0.9.0
ucimlrepo==0.0.7
